Fix eventual consistency bug in ec2_group
The aws api will occasionally throw an 'InvalidGroup.NotFound' error
on the ec2_group module's final call to get_all_security_groups().

This is presumed to happen because the previous call went to an api
endpoint that had the SG information propagated to it, while the
final call went to a different endpoint that had yet to receive it.

This commit uses the data returned from first successful check, instead
of re-querying the API after a success. It also implements an
exponential retry backoff, as recommended by AWS.